requirements.txt
click==7.1.1
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==1.0.0
gunicorn==20.0.4
mysql-connector==2.2.9